Add stroage_address_try_from_felt.
