+++ corelib/src/starknet.cairo
@@ -17,6 +17,7 @@ use storage_access::storage_read_syscall;
 use storage_access::storage_write_syscall;
 use storage_access::storage_address_from_base;
 use storage_access::storage_address_from_base_and_offset;
+use storage_access::storage_address_try_from_felt;
 
 // ContractAddress
 mod contract_address;
