+++ crates/cairo-lang-sierra-ap-change/src/core_libfunc_ap_change.rs
@@ -247,7 +247,9 @@ pub fn core_libfunc_ap_change<InfoProvider: InvocationApChangeInfoProvider>(
         CoreConcreteLibfunc::Pedersen(_) => vec![ApChange::Known(0)],
         CoreConcreteLibfunc::StarkNet(libfunc) => match libfunc {
             StarkNetConcreteLibfunc::ContractAddressConst(_) => vec![ApChange::Known(0)],
-            StarkNetConcreteLibfunc::ContractAddressTryFromFelt(_) => {
+
+            StarkNetConcreteLibfunc::ContractAddressTryFromFelt(_)
+            | StarkNetConcreteLibfunc::StorageAddressTryFromFelt(_) => {
                 vec![ApChange::Known(5), ApChange::Known(6)]
             }
             StarkNetConcreteLibfunc::ContractAddressToFelt(_) => vec![ApChange::Known(0)],
