+++ crates/cairo-lang-sierra-to-casm/src/invocations/starknet/interoperability.rs
@@ -1,14 +1,12 @@
-use cairo_felt::Felt;
 use cairo_lang_casm::builder::CasmBuilder;
 use cairo_lang_casm::casm_build_extend;
 use cairo_lang_casm::cell_expression::CellExpression;
 use cairo_lang_sierra::extensions::consts::SignatureAndConstConcreteLibfunc;
 use cairo_lang_sierra_gas::core_libfunc_cost::SYSTEM_CALL_COST;
-use num_bigint::{BigInt, ToBigInt};
+use num_bigint::BigInt;
 use num_traits::Signed;
 
 use super::{CompiledInvocation, CompiledInvocationBuilder, InvocationError};
-use crate::invocations::misc::validate_under_limit;
 use crate::invocations::{
     add_input_variables, get_non_fallthrough_statement_id, CostValidationInfo,
 };
@@ -86,51 +84,5 @@ pub fn build_contract_address_const(
 
     Ok(builder.build_only_reference_changes(
         [ReferenceExpression::from_cell(CellExpression::Immediate(libfunc.c.clone()))].into_iter(),
-    ))
-}
-
-/// Handles the contract_address_try_from_felt libfunc.
-pub fn build_contract_address_try_from_felt(
-    builder: CompiledInvocationBuilder<'_>,
-) -> Result<CompiledInvocation, InvocationError> {
-    let addr_bound: BigInt = BigInt::from(1) << 251;
-    let [range_check, value] = builder.try_get_single_cells()?;
-    let failure_handle_statement_id = get_non_fallthrough_statement_id(&builder);
-    let mut casm_builder = CasmBuilder::default();
-    add_input_variables! {casm_builder,
-        buffer(2) range_check;
-        deref value;
-    };
-    let auxiliary_vars: [_; 4] = std::array::from_fn(|_| casm_builder.alloc_var(false));
-    casm_build_extend! {casm_builder,
-        const limit = addr_bound.clone();
-        let orig_range_check = range_check;
-        tempvar is_valid_address;
-        hint TestLessThan {lhs: value, rhs: limit} into {dst: is_valid_address};
-        jump IsValidAddress if is_valid_address != 0;
-        tempvar shifted_value = value - limit;
-    }
-    validate_under_limit::<1>(
-        &mut casm_builder,
-        &(-Felt::from(addr_bound.clone())).to_biguint().to_bigint().unwrap(),
-        shifted_value,
-        range_check,
-        &auxiliary_vars,
-    );
-    casm_build_extend! {casm_builder,
-        jump Failure;
-        IsValidAddress:
-    };
-    validate_under_limit::<1>(&mut casm_builder, &addr_bound, value, range_check, &auxiliary_vars);
-    Ok(builder.build_from_casm_builder(
-        casm_builder,
-        [
-            ("Fallthrough", &[&[range_check], &[value]], None),
-            ("Failure", &[&[range_check]], Some(failure_handle_statement_id)),
-        ],
-        CostValidationInfo {
-            range_check_info: Some((orig_range_check, range_check)),
-            extra_costs: None,
-        },
     ))
 }
